# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AegisOps AI
  description: AI-powered autonomous SRE platform with anomaly detection, incident
    investigation, and human-approved remediation.
  version: 1.0.0
paths:
...
  /api/telemetry/logs:
    get:
      tags:
      - Telemetry & Observability
      summary: Get Logs
      operationId: get_logs_api_telemetry_logs_get
      parameters:
      - name: service_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Service Name
      - name: level
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Level
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LogEntry'
                title: Response Get Logs Api Telemetry Logs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/telemetry/time-travel:
    get:
      tags:
      - Telemetry & Observability
      summary: Get Time Travel Snapshot
      description: 'Replays infrastructure state at any point in time T - minutes_ago.

        Allows scrubbing through past telemetry and service health.'
      operationId: get_time_travel_snapshot_api_telemetry_time_travel_get
      parameters:
      - name: minutes_ago
        in: query
        required: false
        schema:
          type: integer
          maximum: 120
          minimum: 0
          default: 5
          title: Minutes Ago
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LogEntry:
      properties:
        id:
          type: integer
          title: Id
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        service_name:
          type: string
          title: Service Name
        level:
          type: string
          title: Level
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
        path:
          anyOf:
          - type: string
          - type: 'null'
          title: Path
        message:
          type: string
          title: Message
        metadata_json:
          additionalProperties: true
          type: object
          title: Metadata Json
          default: {}
      type: object
      required:
      - id
      - timestamp
      - service_name
      - level
      - message
      title: LogEntry
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer